# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api"
        }
    ],
    "paths": {
...
        "/worklog/start": {
            "post": {
                "summary": "Start a new work log",
                "operationId": "02bd1842f32c90f562eef36abf8376a2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "project_id",
                                    "start_time"
                                ],
                                "properties": {
                                    "project_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "start_time": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2023-01-01T10:00:00Z"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Work log started",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WorkLog"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/worklog/{id}/pause": {
            "post": {
                "summary": "Pause an existing work log",
                "operationId": "2ebe0898c5aa235878e7e199daa8e34f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "pause_time",
                                    "keyboard_activity",
                                    "mouse_activity"
                                ],
                                "properties": {
                                    "pause_time": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2023-01-01T11:00:00Z"
                                    },
                                    "keyboard_activity": {
                                        "type": "integer",
                                        "example": 50
                                    },
                                    "mouse_activity": {
                                        "type": "integer",
                                        "example": 60
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Work log paused",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WorkLog"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Work log not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Work log not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/worklog/{id}/resume": {
            "post": {
                "summary": "Resume a paused work log",
                "operationId": "3ae8b0f8491d5afc7862b023bdecce02",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "resume_time"
                                ],
                                "properties": {
                                    "resume_time": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2023-01-01T11:30:00Z"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Work log resumed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WorkLog"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Work log not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Work log not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/worklog/{id}/stop": {
            "post": {
                "summary": "Stop a work log",
                "operationId": "e957dcb64f5b1df0129b4d0566030335",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "end_time",
                                    "keyboard_activity",
                                    "mouse_activity"
                                ],
                                "properties": {
                                    "end_time": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2023-01-01T12:00:00Z"
                                    },
                                    "keyboard_activity": {
                                        "type": "integer",
                                        "example": 70
                                    },
                                    "mouse_activity": {
                                        "type": "integer",
                                        "example": 80
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Time tracking stopped",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Time tracking stopped"
                                        },
                                        "log": {
                                            "$ref": "#/components/schemas/WorkLog"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Work log not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Work log not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/worklog/screenshot": {
            "post": {
                "summary": "Upload a screenshot for a work log",
                "operationId": "3c91d083b1dd589b316720b6f63cbd0c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "worklog_id",
                                    "screenshot"
                                ],
                                "properties": {
                                    "worklog_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "screenshot": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Screenshot uploaded successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Screenshot"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "WorkLog": {
                "required": [
                    "id",
                    "user_id",
                    "project_id",
                    "start_time",
                    "hours_logged"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "project_id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "start_time": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T10:00:00Z"
                    },
                    "hours_logged": {
                        "type": "number",
                        "format": "float",
                        "example": 2.5
                    },
                    "keyboard_activity": {
                        "type": "integer",
                        "format": "int32",
                        "example": 120
                    },
                    "mouse_activity": {
                        "type": "integer",
                        "format": "int32",
                        "example": 150
                    },
                    "end_time": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T12:00:00Z"
                    }
                },
                "type": "object"
            },
            "Screenshot": {
                "required": [
                    "id",
                    "work_log_id",
                    "screenshot_path",
                    "captured_at"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "work_log_id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "screenshot_path": {
                        "type": "string",
                        "example": "path/to/screenshot.jpg"
                    },
                    "captured_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T10:00:00Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": {
        "bearerAuth": {
            "type": "http",
            "scheme": "bearer",
            "bearerFormat": "JWT"
        }
    }
}